Require lvm with  ignored_mdas fix
PVs blonging to a VG are reported as free if the PV is listed
before reading the VG metadata, which is in another PV.
This may cause false getVGInfo() reports.
The present lvm version fixes it.

Note that the Fedora required version is not updated as lvm2
package is not updated there.

Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=798635
Related to: https://bugzilla.redhat.com/show_bug.cgi?id=836663

Change-Id: I9fdf09a2637f096201f094918654e3f52663bc2d
Signed-off-by: Eduardo <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8636
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>