DeferableContext: Change to RollbackContext for easy undoing
We often need to perform a series of operations:
    op[0], op[1], ... op[N]
These operations may allocate files, locks, connections, and op[K] may
depend on op[K-1] 's result

Sometimes it's not feasible to create context manager for each resource,
because the number of the resource involved in a transaction can be a
variable, for example, reading from a configuration file, but there is no
way to use a variable number of the "with" statement, and
contextlib.nested is being deprecated. So we need a concise framework
to do rollback.

This patch subclass DeferableContext and proposes an idiom to do this
kind of rollback. DeferableContext re-raises the last exception, while
this patch re-raises the first exception, for the earliest exception
may be the root cause and most helpful when investigate the problem.

Change-Id: Ibc932637dd81c3becf92de34ea647c1cea136111
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/8671
Reviewed-by: Dan Kenigsberg <danken@redhat.com>