jsonrpc: Implement the JsonRPC server for the next-gen API
This patch implements a new Binding plugin to serve the next generation vdsm API
over a JsonRPC wire protocol.  The basic format of a message is:

    <size><json-data>

<size> is an unsigned 64 bit integer in big endian format that indicates the
length of <json-data> in bytes.  <json-data> is either a request or a response
in Javascript object notation (JSON).

A request object has the following fields:
    id:  An integer which will be repeated in the matching response
    methodName:  The name of the API method to be called
    args (optional): A JSON object containing arguments to the method

A response object has the following fields:
    id:  An integer which will be the same as the matching request
    result:  The return value of the method (defined in the API schema)
    error:  A JSON object containing error information
            code:  An integer error code
            message:  Error context information

Method calls are dispatched to vdsm using a MethodBridge.  The DynamicBridge
dispatches calls based on an API schema document and a set of schema exceptions.
In this way, the API can be expanded without the need to add more code to this
server infrastructure.  When a request is made, the DynamicBridge attempts to
resolve 'methodName' to a schema defined method.  If found, the necessary
arguments are collected from the request and the function call is dispatched to
the internal vdsm API.

Schema exceptions:
Currently, vdsm does not completely conform to its own API schema.  It is not
possible to correct all of the discrepencies due to the need to maintain
backwards compatibility with current API users.  To facilitate migration to this
API, the DynamicBridge implements a set of overrides to translate between the
schema-defined API and what is implemented in vdsm today.  There are three types
of method overrides and one type of argument override:

Method overrides:
1. Custom call function:  If a new API does not map directly to an existing vdsm
   API (or multiple functions must be called to get the result), a custom 'call'
   function can be defined for the API.  This function will be called instead of
   trying to find a vdsm API to call.

2. Return field name:  Many vdsm functions' return values are nested in a
   dictionary.  This override specifies the key to use when accessing the result
   so that it can be returned un-nested to the caller.

3. Custom result post-processing function:  Some vdsm APIs return results in a
   non-standard format and a special function must be used to reformat the data
   to comply with the schema.

Type override:
Some data types have a different representation internally to vdsm than we have
defined in the schema.  A type override can be used to convert between the two
formats.  Currently, the only kind of translation we do is to rename fields
within a given type.

Change-Id: Idae0faa80ffc6a5af002a8a7151aa40dc9a6673d
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/8614
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>