schema: Return objects rather than UUIDs
Rather than returning UUIDs and forcing the API user to instantiate their own
objects using those UUIDs, just return initialized objects.  This patch just
changes the API schema.  The specific implementation of this idea appears in the
patch that implements the code generator.

This is not exhaustive, but an example of the idea.  The rest can be converted
over time.

Change-Id: I7fdfe69436d65ef8190d8a5f1d941e7a628c77c0
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/7840
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>