Set up better threading in tests to make sure it happens correctly
Some object in VDSM count on better threading being set up. They might
break if it isn't set up or even worse set it up in the middle of a test
run cause weird and unexpected errors in the test run.

Also there might be bugs related to our use of pthreading and they will
not be detected by the tests if it is not set up.

This make sure pthreading is imported and set up before any tests are
run.

Change-Id: Ibc1c7fca8e59feb4f772b2aa5e45e3d6d00b8531
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/8793
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>