storage: Don't ignore nfs_mount_options in vdsm.conf
The nfs_mount_options configuration setting in vdsm.conf was ignored.

If the engine doesn't pass any mount options, and nfs_mount_options are
present (and not the defaults), then use the options from the config
file, as options to 'posixfs' connection. Otherwise, use whatever the
engine passes and/or defaults as was done previously.

Bug-Url: https://bugzilla.redhat.com/826921
Change-Id: I3c1d13bfd98a9b41b9728bdcb7f01b3161c26bd8
Signed-off-by: Greg Padgett <gpadgett@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/5368
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>