image: shrink the new volume When merging block cow volumes
On merge, size of the new volume is currently the sum of virtual
sizes of the original volumes that were merged. With qcow
volumes on block devices we don't use the whole size of the new volume,
so we want to allocate only the actual size needed by the new volume
with some additions to match the vdsm conventions.

Change-Id: I8ace4c48d278cb84ce871bc402643131265c3198
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=887909
Signed-off-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9725
Reviewed-by: Ayal Baron <abaron@redhat.com>