The return of super vdsm tests
We couldn't run svdsm tests because imports didn't exist under vdsm
installation directory when running the tests. We want to be able to run
those tests without having to install vdsm rpm.

Here we add extraPythonPath that included the current directory. It is
useful only for tests, if we run it normally we want vdsm to be
installed in its default directories and we don't need this additional
python path list.

Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Change-Id: I52e935662633601ce85150171c37fbf7c6196c15
Reviewed-on: http://gerrit.ovirt.org/9278
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>