spec: Add MOM as a package dependency
For oVirt 3.2, we are switching to MOM for KSM tuning.  One of the goals of this
release is to ensure that MOM is running in all deployed oVirt hosts (ovirt-node
and standalone).  A recent enough version of MOM is now available upstream for
Fedora and RHEL and we build nightly RPMs on ovirt.org.  I think we are ready to
add this packaging dependency now.

Change-Id: I8e8c5a67066f270b8f19881ac5871ff844dbecaa
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/9272
Reviewed-by: Doron Fediuck <dfediuck@redhat.com>
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>