tests: Support parallel testing in apiTests
The apiTests module tests the JSON-RPC server by creating a server instance on a
specific port and then connecting clients to that port.  Unfortunately, this
means that only one instance of the test can be running on a host at any given
time.  Our Jenkins infrastructure would like to run parallel tests.  To enable
this, we allow the test to find an open port to use.

Change-Id: Icee7a7cac15ba271374b622d4227d5c15e190341
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/9342
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>