Fix block volume atomicity creation issue.
BlockVolume creation is a 2 step process where first an LV is
created and then tags are added to it. If another host refreshes
it's cache between the 2 commands it will have an LV with partial
data on it. To solve this issue, we add an initial tag to the
lvcreate command so that any other host would be able to identify
this volume as incomplete and ignore it.
Special LVs still have not MD tags.

Bug-uri: https://bugzilla.redhat.com/show_bug.cgi?id=876558

Change-Id: I40cd67e563935de663d938cbc1bc9cf152802448
Signed-off-by: Eduardo <ewarszaw@redhat.com>
Reviewed-on: https://gerrit.eng.lab.tlv.redhat.com/3299
Reviewed-by: Ayal Baron <abaron@redhat.com>
Tested-by: Haim Ateya <hateya@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9358
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>