iscsi.py: fix log message format
The logging methods accept the first argument as message, and rest
arguments as format values and options. The original code wrongly splits
a message into several arguments, so only the first part is recognized as
message and the rest are recognized as values.
Since there are no format specifiers in the actually recognized message,
it will raise error when it finds there are format values.

This patch utilizes the implicit string concat feature and fix this bug.

Change-Id: I005f6343f2d8775dbf8406dffe04f6de93a81fce
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/9690
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Tested-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>