sdcache: add refresh to connectStoragePool
When connecting to a storage pool we should clear the domain cache to
pick up all the changes that might have been introduced by other hosts.
Additionally since StoragePool.refresh has a similar behavior to
connectStoragePool it should also contain the invalidateStorage call to
force an iscsi rescan.

In this patch:
* reinstate the sdCache.refresh call in connectStoragePool
* add invalidateStorage to StoragePool.refresh

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=879253
Change-Id: I2d3adcff7bb0e97be5c797cd720c6353920d9db0
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9422
Reviewed-by: Ayal Baron <abaron@redhat.com>
Tested-by: Daniel Paikov <paikov@gmail.com>