Add a flag to libvirtvm's status to report watchdog event
This flag will be set once watchdog event is triggered.
The client can get watchdog event by checking the 'watchdogEvent'
key in the return value of getVmStats or getAllVmStats API.
This flag will be overwriting if watchdog fires more than once for
the same VM before it was polled.
And the flag will be not cleanup after it is set unless the client
restart or shutdown VM.
So add a timestamp in the message of this flag.

The timestamp is the time in seconds since the Epoch.

When the client poll this flag, it should record and check the
timestamp to make sure that the event is already triggered.

Change-Id: I7aa676299c748d761e216469569156140a79af65
Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/9429
Reviewed-by: Doron Fediuck <dfediuck@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>