a more delicate rollback
If things go wrong during setupNetwork, we try to revert the networking
state to what it used to be. Until this patch, we stopped the network
service, reverted all ifcfg-* files, and restarted networking.

This procedure disruppted all connection, even those unrelated to the
ones being set up.

With this change, we are taking down only affected devices, and revert
them to their pre-setupNetwork state.

Change-Id: If413164a34a1e6f0d7e4ef75ba931e630a26e666
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9506
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Tested-by: Meni Yakove <myakove@redhat.com>