libvirtvm: prepare for <interface type='network'>
In the near future (actually, in the following patch), I would like to
define virtual interfaces with <source network>. If such a VM is
migrated to a host expecting <source bridge>, bad thing would happen.

This patch extracts <interface>'s network for bridge- and network-based
interfaces alike. It should be applied to all the hosts in the cluster
before generating the first network-based interface.

Change-Id: Ib1eeb8c008d95c8812d855c7e073b5f78ffd67b2
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9531
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Hunt Xu <mhuntxu@gmail.com>