Add default MTU=1500 to NIC's ifcfg files.
When we add bridge with custom MTU on top of nic, this MTU will be set properly
on interface. But if we'll remove this bridge (including removing of MTU
keyword from NIC's ifcfg file) and add instead the bridge with default MTU
(without MTU keyword at all in ifcfg file) the actual MTU on the interface
will stay as before.
The only way to change the actual MTU is add MTU keyword explicitly to ifcfg file.

Change-Id: I0553a3956ea06586e0ece758932f00a9af640611
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=873745
Signed-off-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9547
Reviewed-by: Hunt Xu <mhuntxu@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>