Have direct file not use the O_DIRECT flag on tmpfs and ramfs
tmpfs and ramfs don't support the O_DIRECT flags. This is intentional as
O_DIRECT tells the kernel to bypass the page-cache and those file
systems live solely on the page cache. Since the effect desired by
direct IO is accomplished on these file systems without the flag there
is no reason for the use to test the FS every time before opening a file
for direct access.

The reason we keep the same class instead of falling back to the regular
file object is so we keep the semantics of the DirectFile() class and
the user doesn't have to care that the underlying FS doesn't really
support direct IO.

Change-Id: I7db4136c1a34d960b17312c2c785fc3234b24b92
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9595
Reviewed-by: Shu Ming <shuming@linux.vnet.ibm.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>