Adding monkeyPatch that replaces _start function of superVdsmProxy for tests
This because changes in svdsm internal parameters can cause unexcpected
behaviour and we can avoid it, at least when adding python path. This
implementation is only useful for tests and revert changes that cause
svdsm to crash.
Apparently when adding empty PYTHONPATH in front of sude in normal run
svdsm can't initialize itself and die, instead of explaining the reason
I'm trying to avoid changing vdsm implementation and modify only the
tests.

Change-Id: I94668b165fa858221899698f0679fff8eb5e2223
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9643
Reviewed-by: Dan Kenigsberg <danken@redhat.com>