vdsmd.init: drop check_port_taken
net-tools is not installed on F18 by default, and was obsoleted by
iproute a decade ago. This is a good opportunity to drop the whole
check_port_taken attempt from the daemon startup. This check is
inherently raceful, it was applied only to the the xmlrpc port (even if
jsonrpc is used), and is utterly unhelpful if we're using systemd (that
reports error to system log, just like vdsm.log).

Change-Id: I8417ae3bb2f09404787cd8915984aaa6fb779e7a
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9651
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>