functional tests: Wait till backend server of storage connection reference is connected
Storage connection reference performs the actual connecting in
asynchronous manner, so add a retryAssert on the storage connection
status to wait it gets connected. Otherwise the creation of storage
domain will fail if the backend connection in VDSM is not ready.

Extract a super class for backend servers in functional test, the
waiting for the ready of the connection is put in the super class,
because this waiting can be useful to iscsi and glusterfs testing.

Change-Id: I466b8e8cbfe6e1b04f8ad902f54fa1177b746feb
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/10352
Reviewed-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>