mount.py: handle (deleted) mount points
If an NFS mount is removed from the server, client side kernel appends
" (deleted)" to it specification. It comes from fs/dcache.c's d_path()
function:

  Convert a dentry into an ASCII path name. If the entry has been deleted
  the string " (deleted)" is appended. Note that this is ambiguous.

In this patch we assume that every mount specification ending with this
string is indeed a deleted one, and not a funky nfs server provided by
an evil admin that hates Linux.

Change-Id: I8fdeda5f2900d818912ef5a939b60687462d5344
Bug-Url: http://bugzilla.redhat.com/807351
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/3261
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Tested-by: Haim Ateya <hateya@redhat.com>