jsonrpc: Basic AMQP 1.0 transport support
I don't set the port on purpose, the configuration scheme will have to
be changed to more appropriately support multiple transports. This will
be done in another patch.

Since qpid-proton is in package review for fedora I don't add it as a
requirement in the RPM or for testing. It will be used if present in the
system.

If you know how to review packages for fedora you are welcome to visit
the BZ[1].

We are planning on moving the binding files away from vdsm_api to vdsm
code so that only client side and API code is in the vdsm_api dir.
When that is done we will need to create an RPM for each transport so
that users don't have to install dependencies for all transports.

[1] https://bugzilla.redhat.com/874105

Change-Id: I850005e7375472bbf6238fd38b10d8f1b5e2a191
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/9987
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>