tests/functional: Use deleteVolume instead of deleteImage
Since we are creating a volume, the undo/rollback should
do deleteVolume instead of deleteImage. deleteImage does
not work well for derived filedomains (eg. glusterSD) as
it does not honor the volumeClass associated with the
derived domain, so its not possible to write a test using
deleteImage for glusterSD. Technically too, deleteVolume
fits the current flow of testcase better than deleteImage
as there isn't a chain of volumes being created to be
really justified to use deleteImage

Change-Id: Id9b8f598fcad03b689e14d09ab33c98ec1626f3d
Signed-off-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/10195
Reviewed-by: Dan Kenigsberg <danken@redhat.com>