storage connection monitor: don't recover an unmanaged connection
A connection can be deleted (unmanage) when the storage connection
monitor thread is in in Event.wait(). When the monitor thread continues
to execute, it thinks the connection is lost and try to recover it, but
the fact is it is unmanaged. This bug prevent iscsi connections from
releasing cleanly.

This patch checks if the connection is still managed by the monitor
before recovering it.

Change-Id: I9bc2b1f18f3421039b1d85d1e7a47c746ebfc31a
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/10297
Reviewed-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>