net: drop nicSort
nicSort was introduced so that vdsm calls ifup on bonded nics in the
same order that initscripts does it, in order to obtain the same master
mac address.

A simpler and safer solution is to delegate ifup of bonded nics to
initscripts, which does that when we ifup the bond device.

In this commit we drop nicSort together with its remaining usage.

Change-Id: If8fdbb281030ab7f0eec38481045dbd6c2603ce1
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/10585
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Tested-by: Meni Yakove <myakove@redhat.com>