Don't crash on libvirt network re-definition.
This patch is aimed to solve the libvirt syncing bug described in
bugurl. In short, the vdsmd restart scenario is solved by the
previous patch (30c63c5).

This one solves the case when the interface backing a bridge/network
has disappeared (and thus is not reported by netinfo to the engine)
and then the engine sends a new setupnetwork that attempts to
recreate the partially defined network. Up until now, libvirt would
report an exception. Now we make a best-effort to remove the network
before trying to add it again, by adding a control case in the
setupNetworks logic where networks to be removed are usually removed.

Bug-Url: https://bugzilla.redhat.com/861701
Change-Id: I76d95e20b7aa99280e604abdb1663c6c5c7dd32e
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/10901
Tested-by: oVirt Jenkins CI Server
Reviewed-by: Dan Kenigsberg <danken@redhat.com>