Simplify netinfo get.
This patch refactors netinfo to have the network data fetching on
a separate function. This reduces the get() complexity and will
allow the possibility of reusing the new method (via a future patch
wrapping of it) for more fine-grained information retrieval, e.g.,
update just a network info of a Netinfo instance.

Change-Id: I5c3e617d10308d0bfa4deaed329b2e10ef54077c
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/11080
Reviewed-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-by: Igor Lvovsky <ilvovsky@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>