hooks: Add OpenStack Network vNIC hook
This hook is used to enable OS-network agent to connect a vNIC as it
wants, instead of trying to connect to the bridge that is sent by engine
(which doesn't exist, as the vNIC should be managed by the agent).

Upon migration, the updated XML is being sent, so no further hook
besides this one is required.

Change-Id: I744b83d9c6027bd817e5d4171a77a005611b9818
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/11108
Reviewed-by: Dan Kenigsberg <danken@redhat.com>