vdsm: free loop device when umounting floppy in mkimage.py
After umounting floppy successfully, it doesn't free automatically
the loop device. Here, we add the support to free the loop device
in the umount function of Mount class and free it to prevent it
using up all loop devices.

The issue only happened in the special case that node is built using
livecd-tools(version: 16.8-1) and rpms from centos63. Because:
1. Centos63 disables auto-destruction feature and still depends on
   /etc/mtab to free the loop device.
2. Livecd-tools links /etc/mtab to /proc/self/mounts, so umount can't
   get suggestion to free the loop device from /etc/mtab.

Change-Id: Ia02cfaf4d8ba46de6d42b12d15351d3ab1cb3d03
Signed-off-by: jarod.w <work.iec23801@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/11132
Reviewed-by: Hunt Xu <mhuntxu@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>