Copy truncate vdsm-reg.log
When the vdsm-reg program doesn't stop (if it fails to contact the
engine, for example) it will continuously generate messages that can
flood the log. Moving the log file is not enough because the program
keeps it open, so the space is not freed. As the program opens the file
for appending it is safe to copy and truncate the file.

Change-Id: I92bea487a70fa2f39a88be417faa4efd7bb99f08
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/11282
Reviewed-by: humble devassy <hchiramm@redhat.com>
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>