misc: handle properly timeout=-1 in NoIntrPoll
The parameter timeout=-1 (no timeout) of NoIntrPoll wasn't properly
handled by the previous implementation; it was in fact returning an
empty set of file descriptors if the process was interrupted by a
signal. This patch ensures that the parameter is taken in account
(do not return until there is an event) even when the process is
interrupted.

Change-Id: Iefc4d17559d3335ef6699cb83923c3bd255c916b
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/11394
Reviewed-by: Dan Kenigsberg <danken@redhat.com>