tests: Make getFreePort work in a multithreaded environment
getFreePort does not currently work reliably if multiple copies are running
concurrently.  Since each copy starts at the same port, the chance of choosing
the same port number at the same time is great.  Switch to letting the OS find
an available port.

This should fix the intermittent failures of jsonRpcTests and apiTests.

Change-Id: I6adcdde3fd0312ec532b261107af45dbc87f34f3
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/12028
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>