After fail to connect to supervdsm more than 3 time vdsm gets into panic
Due to race between old supervdsm instance to the new instance after
prepareForShutdown, sometimes the socket is removed after
new supervdsm started to listen on it.
_pokeParent thread unlink the socket when distinguish that vdsm is dead.
This can take more time than the time that takes to vdsm to startup and
start the new instance of supervdsm. The unlink removes the socket file
and vdsm cannot communicate with supervdsm.
When the communication fails, vdsm calls panic and restart itself, this
will start supervdsm again as needed.

Change-Id: Iafe112893a76686edd2949d4f40b734646fd74df
Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=910005
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/11932
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/12053