tests: Skip stress tests that behave badly when run in parallel
A particular test: resourceManagerTests.testStressTest works by spawning lots of
threads to test the resource manager.  Unfortunately, when run in parallel with
other tests (as happens on Jenkins), it can cause spurious errors in comletely
unrelated tests by causing the user to exceed its thread limit.  There is no
real way around this problem other than to skip this test when tests can be
running in parallel.

To achieve this, introduce a new nose plugin (similar to SlowTestsPlugin) that
can be activated by a command line parameter or environment variable.  When
active, tests marked as @stresstest will be skipped.

The environment variable NOSE_SKIP_STRESS_TESTS should be added to the Jenkins
unit test job.

Change-Id: Id74717584f0c4753cec58cb1b701e7095a735924
Signed-off-by: Adam Litke <agl@us.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/12173
Reviewed-by: Dan Kenigsberg <danken@redhat.com>