hsm: always check validateNotSPM when disconnecting from a pool
At the moment of this writing vdsm doesn't have a valid flow to
disconnect from a pool if the host is the SPM; the side effect
of calling _disconnectPool (e.g. a failed refreshStoragePool) is
that the SPM resource will remain locked.

If we want to introduce a change to allow the disconnection from
the pool even if the host is the SPM we should do it explicitly
with a specific patch that stops the SPM.

Change-Id: I066d018eb9e5d61422c5d1ab264208cd89b0f951
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/12271
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Vered Volansky <vvolansk@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>