Adding sudo rules to multipath and removing unused rules
When vdsm tries to cp multipath.conf file as part of setupMultipath that
tries to keep the last multipath.conf file into multipath.conf.1, we get an
error that sudo command requires a password.

This patch add rules to allow operations on absolute path of multipath.conf
file.

Removing old code that related to directories that vdsm doesn't use anymore.

Bud-Id: https://bugzilla.redhat.com/show_bug.cgi?id=907389
Change-Id: Ie732400423f12b0a783bd251b4e2cc2e2409ffec
Signed-off-by: Pavel Zhukov <pzhukov@redhat.com>
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/12226
Reviewed-by: Dan Kenigsberg <danken@redhat.com>