Migration warning due to high guest memory changes
Migration process consists in copy guest RAM to new Host. After the
copy, migration code loops back to the beginning and re-copies all pages
modified after the copy. The hope is that the list of pages which must
be copied shrinks with each pass over the memory range.
Guests with large number of memory writes can delay migration process
and even stuck it, if memory writes are faster than memory copy.
This patch add a message to expose this situation, improving the
supportability.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=919201
Change-Id: I5e7c5a2b7e9e97f047fa893a47408a659a40beb2
Signed-off-by: Amador Pahim <apahim@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/12557
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Peter V. Saveliev <peet@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>